requirements.txt
fastapi==0.85.1
pydantic==1.10.2
python-dotenv==0.21.0
starlette==0.20.4
uvicorn==0.19.0
gunicorn==20.1.0
pydantic[email]
pymongo==4.3.3
python-multipart==0.0.6
python-jose[cryptography]
passlib[bcrypt]